requirements.txt
langchain>=0.2.16
langchain-anthropic>=0.3.0
langchain-community>=0.2.16
langgraph>=0.2.9
langsmith>=0.1.117
anthropic>=0.31.1
pydantic>=2.8.2
python-dotenv>=1.0.1
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0